Bug 78293: Extract usePluginFlagsSettled hook and fix timeout reset
Move the plugin flag settling logic from app-contents.tsx into a
dedicated usePluginFlagsSettled hook for better separation of
concerns and testability.

Fix a bug where flagSettlingTimedOut was never reset when
hasPendingPluginFlags became false, which could cause the hook to
skip the timeout on a subsequent pending-flags cycle (e.g. if a
late-loading plugin introduces new pending flags after the initial
batch has settled).
