+++ frontend/public/components/app-contents.tsx
@@ -3,6 +3,7 @@ import { useState, useEffect, useRef, useMemo, Suspense } from 'react';
 import { PageSection } from '@patternfly/react-core';
 import * as _ from 'lodash';
 import { Route, Routes, Navigate, useParams, useLocation, matchRoutes } from 'react-router';
+import { usePluginFlagsSettled } from '@console/app/src/hooks/usePluginFlagsSettled';
 import { usePluginRoutes } from '@console/app/src/hooks/usePluginRoutes';
 import type { Perspective } from '@console/dynamic-plugin-sdk';
 import { useActivePerspective } from '@console/dynamic-plugin-sdk';
@@ -10,7 +11,6 @@ import { getReferenceForModel } from '@console/dynamic-plugin-sdk/src/utils/k8s'
 import { usePluginInfo } from '@console/plugin-sdk/src/api/usePluginInfo';
 import { ErrorBoundaryPage } from '@console/shared/src/components/error/fallbacks/ErrorBoundaryPage';
 import { FLAGS } from '@console/shared/src/constants/common';
-import { useConsoleSelector } from '@console/shared/src/hooks/useConsoleSelector';
 import {
   getPerspectiveVisitedKey,
   usePerspectives,
@@ -159,35 +159,8 @@ const AppContents: FC = () => {
     [pluginInfoEntries],
   );
 
-  const reduxFlags = useConsoleSelector((state) => state.FLAGS);
-
-  // Check if any loaded plugin has extensions with required flags still pending
-  const hasPendingPluginFlags = useMemo(
-    () =>
-      pluginInfoEntries.some((entry) => {
-        if (entry.status === 'pending' || entry.status === 'failed') return false;
-        const extensions = entry.manifest?.extensions;
-        if (!Array.isArray(extensions)) return false;
-        return extensions.some((ext) => {
-          const required = ext.flags?.required;
-          if (!Array.isArray(required) || required.length === 0) return false;
-          return required.some((flagName: string) => flagPending(reduxFlags[flagName]));
-        });
-      }),
-    [pluginInfoEntries, reduxFlags],
-  );
-
-  // Timeout fallback: if flags never resolve (broken hookProvider),
-  // stop blocking after 3 seconds
-  const [flagSettlingTimedOut, setFlagSettlingTimedOut] = useState(false);
-  useEffect(() => {
-    if (allPluginsProcessed && hasPendingPluginFlags && !flagSettlingTimedOut) {
-      const timer = setTimeout(() => setFlagSettlingTimedOut(true), 3000);
-      return () => clearTimeout(timer);
-    }
-  }, [allPluginsProcessed, hasPendingPluginFlags, flagSettlingTimedOut]);
-
-  const showNotFound = allPluginsProcessed && (!hasPendingPluginFlags || flagSettlingTimedOut);
+  const pluginFlagsSettled = usePluginFlagsSettled(pluginInfoEntries);
+  const showNotFound = allPluginsProcessed && pluginFlagsSettled;
 
   const contentRouter = (
     <Routes>
